pyproject.toml
[project]
name = "langgraph-browser-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.45.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.10",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
